+++ src/SectionList.tsx
@@ -23,7 +23,13 @@ const SectionListMemo = React.memo(
     RNSectionList,
     React.PropsWithChildren<SectionListProps<unknown>>
   >((props, passRef) => {
-    return <AnimatedSectionList ref={passRef} {...props} />
+    return (
+      <AnimatedSectionList
+        // @ts-expect-error reanimated types are broken on ref
+        ref={passRef}
+        {...props}
+      />
+    )
   })
 )
 
@@ -42,7 +48,8 @@ function SectionListImpl<R>(
   const ref = useSharedAnimatedRef<RNSectionList<unknown>>(passRef)
 
   const { scrollHandler, enable } = useScrollHandlerY(name)
-  useAfterMountEffect(() => {
+  const onLayout = useAfterMountEffect(rest.onLayout, () => {
+    'worklet'
     // we enable the scroll event after mounting
     // otherwise we get an `onScroll` call with the initial scroll position which can break things
     enable(true)
@@ -104,6 +111,7 @@ function SectionListImpl<R>(
     // @ts-expect-error typescript complains about `unknown` in the memo, it should be T
     <SectionListMemo
       {...rest}
+      onLayout={onLayout}
       ref={ref}
       bouncesZoom={false}
       style={memoStyle}