+++ src/hooks.tsx
@@ -7,8 +7,9 @@ import {
   MutableRefObject,
   useEffect,
   DependencyList,
+  useRef,
 } from 'react'
-import { StyleSheet } from 'react-native'
+import { LayoutChangeEvent, StyleSheet, ViewProps } from 'react-native'
 import { ContainerRef, RefComponent } from 'react-native-collapsible-tab-view'
 import { PagerViewOnPageScrollEvent } from 'react-native-pager-view'
 import Animated, {
@@ -135,20 +136,27 @@ export function useCollapsibleStyle(): CollapsibleStyle {
     containerHeight,
     width,
     allowHeaderOverscroll,
+    minHeaderHeight,
   } = useTabsContext()
   const [containerHeightVal, tabBarHeightVal, headerHeightVal] = [
     useConvertAnimatedToValue(containerHeight),
     useConvertAnimatedToValue(tabBarHeight),
     useConvertAnimatedToValue(headerHeight),
   ]
+
+  const containerHeightWithMinHeader = Math.max(
+    0,
+    (containerHeightVal ?? 0) - minHeaderHeight
+  )
+
   return useMemo(
     () => ({
       style: { width },
       contentContainerStyle: {
         minHeight:
           IS_IOS && !allowHeaderOverscroll
-            ? (containerHeightVal || 0) - (tabBarHeightVal || 0)
-            : (containerHeightVal || 0) + (headerHeightVal || 0),
+            ? containerHeightWithMinHeader - (tabBarHeightVal || 0)
+            : containerHeightWithMinHeader + (headerHeightVal || 0),
         paddingTop:
           IS_IOS && !allowHeaderOverscroll
             ? 0
@@ -163,10 +171,10 @@ export function useCollapsibleStyle(): CollapsibleStyle {
     }),
     [
       allowHeaderOverscroll,
-      containerHeightVal,
       headerHeightVal,
       tabBarHeightVal,
       width,
+      containerHeightWithMinHeader,
     ]
   )
 }
@@ -266,6 +274,7 @@ export const useScrollHandlerY = (name: TabName) => {
 
   const enable = useCallback(
     (toggle: boolean) => {
+      'worklet'
       enabled.value = toggle
     },
     [enabled]
@@ -442,9 +451,9 @@ export const useScrollHandlerY = (name: TabName) => {
   // sync unfocused scenes
   useAnimatedReaction(
     () => {
-      if (!enabled.value) {
-        return false
-      }
+      // if (!enabled.value) {
+      //   return false
+      // }
 
       // if the index is decimal, then we're in between panes
       const isChangingPane = !Number.isInteger(indexDecimal.value)
@@ -457,7 +466,7 @@ export const useScrollHandlerY = (name: TabName) => {
         isSyncNeeded !== wasSyncNeeded &&
         focusedTab.value !== name
       ) {
-        let nextPosition = null
+        let nextPosition: number | null = null
         const focusedScrollY = scrollY.value[Math.round(indexDecimal.value)]
         const tabScrollY = scrollY.value[tabIndex]
         const areEqual = focusedScrollY === tabScrollY
@@ -529,19 +538,59 @@ export function useSharedAnimatedRef<T extends RefComponent>(
   return ref
 }
 
-export function useAfterMountEffect(effect: React.EffectCallback) {
-  const [didExecute, setDidExecute] = useState(false)
-  useEffect(() => {
-    if (didExecute) return
+export function useAfterMountEffect(
+  nextOnLayout: ViewProps['onLayout'],
+  effect: React.EffectCallback
+) {
+  const name = useTabNameContext()
+  const {
+    //tabsMounted,
+    refMap,
+    scrollY,
+    //scrollYCurrent,
+    tabNames,
+  } = useTabsContext()
 
-    const timeout = setTimeout(() => {
-      effect()
-      setDidExecute(true)
-    }, 0)
-    return () => {
-      clearTimeout(timeout)
+  const didExecute = useRef(false)
+  const didMount = useSharedValue(false)
+
+  const scrollTo = useScroller()
+  const ref = name ? refMap[name] : null
+
+  useAnimatedReaction(
+    () => {
+      return didMount.value
+    },
+    (didMount, prevDidMount) => {
+      if (didMount && !prevDidMount) {
+        if (didExecute.current) return
+        if (ref) {
+          const tabIndex = tabNames.value.findIndex((n) => n === name)
+          scrollTo(
+            ref,
+            0,
+            scrollY.value[tabIndex],
+            false,
+            `[${name}] restore scroll position`
+          )
+        }
+        effect()
+        didExecute.current = true
+      }
     }
-  }, [didExecute, effect])
+  )
+
+  const onLayoutOut: NonNullable<ViewProps['onLayout']> = useCallback(
+    (event: LayoutChangeEvent) => {
+      requestAnimationFrame(() => {
+        didMount.value = true
+      })
+      return nextOnLayout?.(event)
+    },
+    [didMount, nextOnLayout]
+  )
+
+  return onLayoutOut
 }
 
 export function useConvertAnimatedToValue<T>(